Release 13
With this release we've switched away from authenticated, encrypted blobs on S3.
For users of prior releases, there is a one-time step required for this process.

For development, you may need to clear the blob symlinks you have locally to
avoid checksum mismatch errors. You can use the following command:

    git clean -d -f -x .final_builds/

Blobs from prior releases are still available in the S3 blobstore. If you need
to switch to a prior release, you'll need to run through the same cleaning
process.