Release 16
This adds TLS support to the syslog ingestor and significantly changes how logstash filters are configured.

Backward Compatibility Break - the `logstash.filters` is no longer an array of files to extract and dynamically compose a set of filters. The value must now be the final concatenated form of the filters you want to use (excluding the wrapping `filter { ... }` parts). Typically you'll automatically update this from a build artifact.

 * BREAKING CHANGE: the `logstash.filters` property now expects a string of the literal filter rules logstash should use (#58)
 * ADD: syslog_tls ingestor support (#59)
 * ADD: drain script for elasticsearch - currently disabled, by default, with the `elasticsearch.drain` property (#51)