Added scp support to the archiver job.
This PR provides another method for archiving your logs. In particular,
when you don't have an outside connection from the network to the
internet, you may want to archive your jobs to another internal location.
It creates a ssh keypair for passwordless ssh and uses the sshpass library
in conjuction with ssh-copy to distribute the public key to the archive machine
without user interaction. Once the key is transferred, the public key is
removed from the archiver job vm.